requirements.txt
deskew==1.3.3
numpy==1.22.3
opencv_python==4.5.5.64
Pillow==9.2.0
pytesseract==0.3.10
scikit_image==0.19.3
